Initial commit
This commit includes an initial implementation
of an AHK script that will interface with the Kenku Remote REST API.

I have not tested it at all, or even tried to execute it.

It includes a few simple hotkeys that can be used to play/pause the current
track, or jump to the next/previous track.